fix the automated unit tests for server routes, by separating the server startup from the "app" code
